streamlit/requirements.txt
streamlit==1.38.0
pandas==2.2.2
plotly==5.24.1
kafka-python==2.0.2